Fix SRT time rounding bug and add tests
